[refs #40] Refactored typography to work like GDS
Since we are using GDS (Government Digital Service) grid and spacing,
it made sense to implement their typography code to build upon.
A Sass map with font values is in settings. The functions and mixins
are within tools. The CSS output (typography classes) are within styles.
The 'overrides' code is within utilities.


Former-commit-id: 565ce4d25c173a104e3d0af7b2dc9cc134ebc5c3