[refs #306] Back link, button, footer and header
Back link - amended code to include a surrounding div in order
so it could be displayed as a block element.
Having the `<a>` set as `display: block;`
caused issues with the hover styling.

Button - changed the font-size to scale on mobile.

Footer - amended the visited link colour to not turn purple.
Updated the logo sizes with the new mixin `nhsuk-logo-size`.

Header - we now have 7 variations of the header.
Each one is documented in the readme and has an example in the docs.
This includes variants for transactional services and also for services
that are not transactional and need the search (eg: Service manual).


Former-commit-id: 066072c0eac6a6b87c45a178797d2c14292d5432