[refs #320] Remove search suggestions heading
It is not currently possible to have the search suggestions heading at
the top of the dropdown list so this has been removed.

Fixed an IE<=11 issue with '.remove()' not supported.

Removed the bolding of the typed text because VoiceOver pronounced the
search suggestion in a funny way. For example, if you start typing 'he'
the results may include 'health', 'behind the headlines', 'heart'. These
would be pronounced as
'he' 'alth',
'behind t' 'he' 'he' 'adlines'
'he' 'art'
which sounds weird and not really acceptable.


Former-commit-id: 207529b2ce78c624a50be7d5a5476430eaf668af