[refs #177] Remove <time> element
Removed the <time> element around the review dates because it fails
validation without a datetime attribute, which cannot be automatically
generated (yet) from the date value. Even though datetime attribute is
not strictly necessary, this can only be the case if the review date
entered is a valid date or time string, which cannot be guaranteed.

More info here:
https://www.w3.org/TR/2011/WD-html5-author-20110705/the-time-element.html

Former-commit-id: 556fdcfb965cd48b510f0b04c23c3375ca313d8b