Stop compiling markdown files to HTML within the docs
We explored the idea of compiling markdown files, such as the README.md,
into HTML files so that they could be viewed within the browser. The
service manual is being used for documentation instead of the GitHub pages.
Remove any files or gulp packages related to the compiling of markdown
files.


Former-commit-id: 07ef83c081aa742808da9c511dc30554ae90a95c