Broaden the main app content width
When we originally ported over the NHS Service Manual code into this repository we excluded some bits, like the custom header and custom container width styling, so we could lean on the defaults on the design system. This resulted in a smaller content width on our docs site than what's on the NHS Service Manual.

As Ciaran and Gary have noticed, this reduced width meant that certain design examples that rely on responsive behaviour would show smaller breakpoint versions on typical laptop screens.

This change attempts to fix this by increasing the app content width to `1100px` (as well as the header) to match the same width used on the NHS Service Manual.